Increased DB test coverage to 100%. Needed a rename of every package, so merging will be painful...
