+++ server/src/main/java/uk/ac/cam/cl/echo/extrusionfinder/server/configuration/Configuration.java
@@ -1,8 +1,10 @@
-package uk.ac.cam.echo.extrusionfinder.configuration;
+package uk.ac.cam.cl.echo.extrusionfinder.server.configuration;
 
 public class Configuration {
     public static final String MONGO_HOST = "localhost";
     public static final int MONGO_PORT = 27017;
 
     public static final String ZERNIKE_MAP_ID = "zernike";
+
+    private Configuration(){}
 }
