+++ server/src/main/java/uk/ac/cam/cl/echo/extrusionfinder/server/database/MongoDBCollectionManager.java
@@ -1,4 +1,4 @@
-package uk.ac.cam.echo.extrusionfinder.database;
+package uk.ac.cam.cl.echo.extrusionfinder.server.database;
 
 import com.mongodb.BasicDBObject;
 import org.mongojack.JacksonDBCollection;
