+++ server/src/main/java/uk/ac/cam/cl/echo/extrusionfinder/server/database/MongoInstance.java
@@ -1,8 +1,8 @@
-package uk.ac.cam.echo.extrusionfinder.database;
+package uk.ac.cam.cl.echo.extrusionfinder.server.database;
 
 import com.mongodb.DB;
 import com.mongodb.MongoClient;
-import uk.ac.cam.echo.extrusionfinder.configuration.Configuration;
+import uk.ac.cam.cl.echo.extrusionfinder.server.configuration.Configuration;
 
 import java.net.UnknownHostException;
 import java.util.HashMap;
@@ -15,6 +15,8 @@ import java.util.Map;
  */
 class MongoInstance {
     private static Map<String, DB> databases = new HashMap<>();
+
+    private MongoInstance(){}
 
     /**
      * @return                      Handle to a MongoDB database
